+++ changelog.md
@@ -22,4 +22,4 @@
 
 ## TBD
 - Add `Vector::negative`
-- Add `Radians::from_degree`
+- Add `Radians::try_from_degrees`